optimize user control
