+++ .env.example
@@ -2,5 +2,5 @@ export openai_apikey=
 export bot_token=
 export api_endpoint=https://api.openai.com/v1/chat/completions
 export default_api_endpoint=https://api.openai.com/v1/chat/completions
-export model=gpt-3.5-turbo-0301
+export model=gpt-3.5-turbo
 export DEVELOPER_CHAT_ID=
