+++ src/utils.py
@@ -13,21 +13,16 @@ from constants.messages import (
     APPROVED_MESSAGE,
     DECLINE_MESSAGE,
     )
-from allowed import add, not_allowd, quota_exceeded
+from allowed import add
 
 
 async def waring(update: Update, context: ContextTypes.DEFAULT_TYPE, msg) -> None:
-    if msg == not_allowd:
-        await update.message.reply_text(
-            text=NOT_PERMITED.format(user_id=context._user_id),
-            pool_timeout=3600.0
-        )
-        await apply_to_prove(update, context)
-        return
-
-    if msg == quota_exceeded:
-        await update.message.reply_text(text=msg)
-        return
+    await update.message.reply_text(
+        text=NOT_PERMITED.format(user_id=context._user_id),
+        pool_timeout=3600.0
+    )
+    await apply_to_prove(update, context)
+    return
 
 
 async def apply_to_prove(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
@@ -36,7 +31,7 @@ async def apply_to_prove(update: Update, context: ContextTypes.DEFAULT_TYPE) -> 
     markup = ReplyKeyboardMarkup(
         yes_or_no_keyboard, resize_keyboard=True, one_time_keyboard=True
     )
-    await bot.sendMessage(
+    message = await bot.sendMessage(
         chat_id=os.getenv("DEVELOPER_CHAT_ID", 0),
         text=ASK_FOR_PERMITED.format(name=update.message.chat.first_name, user_id=context._user_id) ,
         reply_markup=markup,
@@ -55,12 +50,17 @@ async def apply_to_prove(update: Update, context: ContextTypes.DEFAULT_TYPE) -> 
         return
 
     if isinstance(updater, Iterable) and len(updater) > 0:
+        # import pdb
+        # pdb.set_trace()
         reply_obj = updater[0]
         permit_message = reply_obj.message
         if permit_message.chat.id == int(os.getenv("DEVELOPER_CHAT_ID", 0)) and permit_message.text == "Approved":
-            _ = add(context._user_id, update.message.chat.first_name)
+            _ = add(context._user_id, update.message.chat.first_name, 1)
             await update.message.reply_text(text=APPROVED_MESSAGE)
+            await message.edit_reply_markup(reply_markup=None)
             return
         if permit_message.chat.id == int(os.getenv("DEVELOPER_CHAT_ID", 0)) and permit_message.text == "Decline":
+            _ = add(context._user_id, update.message.chat.first_name, 0)
             await update.message.reply_text(text=DECLINE_MESSAGE)
+            await message.edit_reply_markup(reply_markup=None)
             return
